requirements.txt
# Core dependencies
streamlit==1.28.0
python-dotenv==1.0.0

# LLM providers
openai==1.3.0
groq==0.4.1
google-generativeai==0.3.0

# Embeddings and Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Document processing
PyPDF2==3.0.1

# Web search
requests==2.31.0

# Additional utilities
torch==2.1.0
transformers==4.35.0
